complete all static checks
