+++ gwcli/stylesheet/colorizer/colorizer.go
@@ -6,7 +6,7 @@
  * BSD 2-clause license. See the LICENSE file for details.
  **************************************************************************/
 
-// colorizer provides common utilities that rely on the stylesheet.
+// Package colorizer provides common utilities that rely on the stylesheet.
 // These functions are some degree of a mish-mash, but all are in the vein of supporting a
 // consistent UI. Any UI elements that look similar across actions should go here as well as
 // functional applications of the rest of the stylesheet.
@@ -24,12 +24,12 @@ import (
 	"github.com/spf13/cobra"
 )
 
-// tea.Printf wrapper that colors the output as an error
+// ErrPrintf is a tea.Printf wrapper that colors the output as an error.
 func ErrPrintf(format string, a ...interface{}) tea.Cmd {
 	return tea.Printf("%s", stylesheet.ErrStyle.Render(fmt.Sprintf(format, a...)))
 }
 
-// Given a command, returns its name appropriately colored by its group (action or nav).
+// ColorCommandName returns the given command's name appropriately colored by its group (action or nav).
 // Defaults to nav color.
 func ColorCommandName(c *cobra.Command) string {
 	if action.Is(c) {
@@ -39,8 +39,7 @@ func ColorCommandName(c *cobra.Command) string {
 	}
 }
 
-// if field == selected, returns the selection rune.
-// otherwise, returns a space.
+// Pip returns the selection rune if field == selected, otherwise it returns a space.
 func Pip(selected, field uint) string {
 	if selected == field {
 		return lipgloss.NewStyle().Foreground(stylesheet.AccentColor2).Render(string(stylesheet.SelectionPrefix))
@@ -48,13 +47,13 @@ func Pip(selected, field uint) string {
 	return " "
 }
 
-// Returns a simple checkbox with angled edges.
+// Checkbox returns a simple checkbox with angled edges.
 // If val is true, a check mark will be displayed
 func Checkbox(val bool) string {
 	return box(val, '[', ']')
 }
 
-// Checkbox() but with rounded edges.
+// Radiobox is Checkbox but with rounded edges.
 func Radiobox(val bool) string {
 	return box(val, '(', ')')
 }
@@ -69,16 +68,16 @@ func box(val bool, leftBoundary, rightBoundary rune) string {
 	return fmt.Sprintf("%c%c%c", leftBoundary, c, rightBoundary)
 }
 
-// Displays either the key-bind to submit the action on the current tab or the input error,
-// if one exists, as well as the result string, beneath the submit-string/input-error
+// SubmitString displays either the key-bind to submit the action on the current tab or the input error,
+// if one exists, as well as the result string, beneath the submit-string/input-error.
 func SubmitString(keybind, inputErr, result string, width int) string {
 	alignerSty := lipgloss.NewStyle().
 		PaddingTop(1).
 		AlignHorizontal(lipgloss.Center).
 		Width(width)
 	var (
-		inputErrOrAltEnterColor        = stylesheet.TertiaryColor
-		inputErrOrAltEnterText  string = "Press " + keybind + " to submit"
+		inputErrOrAltEnterColor = stylesheet.TertiaryColor
+		inputErrOrAltEnterText  = "Press " + keybind + " to submit"
 	)
 	if inputErr != "" {
 		inputErrOrAltEnterColor = stylesheet.ErrorColor
@@ -91,7 +90,7 @@ func SubmitString(keybind, inputErr, result string, width int) string {
 	)
 }
 
-// Return the given number, styled as an index number in a list or table.
+// Index returns the given number, styled as an index number in a list or table.
 func Index(i int) string {
 	return stylesheet.IndexStyle.Render(strconv.Itoa(i))
 }
